Detections implemented fully, errors fixed, folders added for modularity and clarity(utils and detections folder), redirecvtions implemented
