+++ lib/utils/routes.dart
@@ -16,6 +16,7 @@ class AppRoutes {
   static const String homepage = "/homepage";
   static const String activity = "/activitypage";
   static const String navbar = "/navbar";
-  static const String camera = "/camera";
-
+  static const String squatdetection = "/squatdetection";
+  static const String pushupdetection = "/pushupdetection";
+  static const String notyetavailable = "/notyetavailable";
 }
